refactor(instrumentation): provider-driven exporter construction and module split
Consolidate exporter construction behind a single ExporterProvider abstraction
(ExporterConfig::Options + create_exporter) and drop the redundant free
create_exporter, so the library OptionsExporterProvider and the callback
provider share one path. Rename Context::observer_with to observer and
try_active to try_new.

Split crates/instrumentation/src/lib.rs into context.rs (Context, backend
runtime), observer.rs (Observer, EventSender, forwarder), and sidecar.rs
(write_sidecar); lib.rs is now module declarations, re-exports, and the e2e
filesystem test. Migrate the codegen bridges, model macro, benches, and
integration tests to the new API.
